issue-994: extract custom command contract/runtime
