+++ Cargo.toml
@@ -23,6 +23,7 @@ members = [
     "crates/tau-memory",
     "crates/tau-github-issues",
     "crates/tau-dashboard",
+    "crates/tau-custom-command",
     "crates/tau-memory",
   "crates/tau-coding-agent",
 ]
