+++ crates/tau-coding-agent/Cargo.toml
@@ -36,6 +36,7 @@ tau-ops = { path = "../tau-ops" }
 tau-memory = { path = "../tau-memory" }
 tau-github-issues = { path = "../tau-github-issues" }
 tau-dashboard = { path = "../tau-dashboard" }
+tau-custom-command = { path = "../tau-custom-command" }
 tau-agent-core = { path = "../tau-agent-core" }
 tau-ai = { path = "../tau-ai" }
 reqwest.workspace = true
