+++ crates/tau-custom-command/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod custom_command_contract;
+pub mod custom_command_runtime;
